requirements.txt
# # FastAPI and ASGI
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# python-multipart==0.0.6

# # Database and Storage
# motor==3.3.2  # MongoDB async driver
# pymongo==4.6.0
# minio==7.2.0  # MinIO S3 client

# # Authentication and Security
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4

# # Data Validation
# pydantic==2.5.0
# pydantic-settings==2.1.0
# email-validator==2.1.0

# # File Processing
# aiofiles==23.2.1
# PyYAML==6.0.1  # YAML parsing
# Pillow==10.1.0  # Image processing
# opencv-python==4.8.1.78  # Computer vision

# # Async and Concurrency
# asyncio==3.4.3

# # Utility
# python-dotenv==1.0.0
# python-dateutil==2.8.2
# httpx==0.25.2
# typing-extensions==4.8.0

# # Data Analytics - YOLO format support
# numpy>=1.26.0
# pandas>=2.1.4  # For data analysis

# # API Documentation
# fastapi[all]  # Complete FastAPI with docs



# # Environment Configuration
# python-decouple==3.8  # Environment variable handling


fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.5.0
minio==7.1.16
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
ultralytics==8.3.228
pyyaml==6.0.1
python-magic==0.4.27
aiofiles==23.2.1
numpy==1.24.0